requirements.txt
pandas>=2.0.0
rapidfuzz>=3.0.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
numpy>=1.24.0
python-dateutil>=2.8.0
jsonschema>=4.0.0
tqdm>=4.65.0
joblib>=1.3.0

